Delete converter-manual directory
The converter-manual directory houses the location of a manual converter that was built in python to convert data from a different Inventory Manager to SIM. The data was stored in an unconventional way within the old Inventory Manager and the converter only works for that specific configuration. It should not be part of the repo.