Redesign totals to be incredibly robust
Before this commit totals in documents didn't really exist. Now, totals
are robust. It should allow almost any combination of values to be
entered, and gives the user full control in to do what they please. This
version of totals will not be the final. The following things still need
to be worked on:
 - Adding and deleting custom expenses / discounts does not work
 - Expenses, rates, and totals are not saved into the database (and as a
   result, recreating the totals column has not been implemented)
Also in this commit, columns names for documents are now accessed from a
struct inside of the CreateDocument class rather than by referencing the
numbers of the columns themselves.