Refactor code for readability and modularity
In short, changes that would take almost an hour of hopping between
convoluted functions can now be done in minutes if not seconds.

In long, The codebase is now much more manageable. Several disgusting
ways of managing tables have been removed. All the placeholder
definitions  in the main.cpp files have been replaced by enums (or in
some cases overhauled to a better method). The codebase is now smaller,
easier to understand, and lighter on memory.

If you squint this almost looks like some good code.