Remove address table and references to it
Addresses prior to this commit were stored in a separate table in the
databse, and references when they were needed to save on disk space.
This system was working, but lead to a very complicated system in order
to actually manage addresses, prevent overwrites, minimize space, etc.
This feature may come back in the future, but the program is much
simpler without it existing.